add TSPulse in tutorial
