+++ TSB_AD/models/README.md
@@ -23,7 +23,7 @@ cd granite-tsfm
 pip install ".[notebooks]"
 ```
 For exact result reproducibility we recommend the use of Python version - 3.12.9, and the IBM granite TSFM release - v0.3.2. 
-
+An example usage is available at [Link](https://github.com/TheDatumOrg/TSB-AD/tree/main/tutorials/TSPulse.py).
 
 If you want to use [Lag-Llama](https://github.com/time-series-foundation-models/lag-llama), please install the following
 ```bash
